157: A Template's Client and Server classes are now resolved through the TemplateResolver.
Previously, RsrService implemented #templateClass, #clientClass, and #serverClass. These methods looked up the Service classes by name. This means that in a use case like Jadeite For Pharo, where the classes may not be in the user's class scope, this would fail. It required the developer to override these methods to get it working -- in addition to implementing a custom TemplateResolver.

Now, all you need to do is implement a custom TemplateResolver. Any restrictions or scope considerations are handled within the custom TemplateResolver.