164: RsrReference subclass lookup no longer uses a Dictionary.
Previously, looking up the RsrReference subclass for an object used a Dictionary that was cached in RsrReference. This causes issue anytime you add a new class as the Dictionary must be initialized again.

Testing in Pharo showed that testing using #== and #ifTrue: was 2.5x faster than the Dictionary lookup. This pattern saves us ~500ms when running the test suite in GemStone.